+++ internal/core/application/input/catalogo_input.go
@@ -3,6 +3,6 @@ package input
 type CatalogoInput struct {
 	Nome          string
 	DuracaoPadrao int
-	Preco         float64
+	Preco         int
 	Categoria     string
 }
